requirements.txt
streamlit
pandas
requests
python-dotenv
openai
langchain
faiss-cpu
langchain_community
tiktoken
langchain-openai